backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.28
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pydantic[email]==2.6.3
pydantic-settings==2.2.1
scikit-learn==1.4.2
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1
websockets==12.0
httpx==0.27.0